Create a separate source set for the shared Blackhole
The CommonBlackhole class is going to be hidden/internalized. Once it is internal, the public Js and WasmJs Blackhole classes won’t be able to inherit from it. This separate source set, shared btw Js and WasmJs, ensures that the targets have a shared implementation, allowing the CommonBlackhole class to be properly internalized or dropped.